refactor: relocate output modules
