+++ src/agent/output/LatexDiffManager.ts
