+++ src/agent/output/StatisticsReporter.ts
