+++ src/agent/output/XmlOutputManager.ts
